% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{applyRoiMask}
\alias{applyRoiMask}
\title{Apply an ImageJ ROI Mask to a TIFF Image}
\usage{
applyRoiMask(roi, frame)
}
\arguments{
\item{roi}{An ImageJ ROI, as an ijroi object.}

\item{frame}{A TIFF image, as a matrix.}
}
\value{
The frame ID, width, height, and pixel values of the masked region, as an integer vector.
}
\description{
Masks a section of a TIFF image corresponding to an ImageJ ROI, including those pixels on the boundary of the ROI.
}
