% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSeparationScores.R
\name{calculateSeparationScores}
\alias{calculateSeparationScores}
\title{Calculate a table containing separation scores for each variable}
\usage{
calculateSeparationScores(
  group1data,
  group2data,
  threshold = 0,
  calculateOptimalThresh = FALSE
)
}
\arguments{
\item{group1data}{Feature table for group 1}

\item{group2data}{Feature table for group 2}

\item{threshold}{Separation threshold, only features achieving separation greater than or equal to this threshold will be output}
}
\value{
A data frame of separation values where column 1 is a list of variable indices, column 2 a list of variable names, and column 3 the corresponding separation scores
}
\description{
This function can be used to calculate the separation score for each variable, where a feature selection threshold can be defined by the user to identify discriminatory variables.
Note that \code{group1data} and \code{group2data} should include only columns of variables (i.e remove cell identifier columns prior to function use)
}
