% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellPhe.R
\docType{package}
\name{CellPhe}
\alias{CellPhe-package}
\alias{CellPhe}
\title{CellPhe: Toolkit for cell phenotyping from time-lapse videos.}
\description{
TODO: Add extended package description.
}
\author{
Laura Wiggins \href{mailto:llw511@york.ac.uk}{llw511@york.ac.uk}
}
