% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{copyFeatures}
\alias{copyFeatures}
\title{Copy features from a pre-existing feature table}
\usage{
copyFeatures(df, minframes)
}
\arguments{
\item{df}{A data frame comprising pre-calculated features (in columns)
for each frame of each cell (in rows). The first column should give the
tracked cell identifier and the second should give the frame number.
All other columns are considered to be features.}

\item{minframes}{The minimum number of frames a cell must be tracked for to
be included in the output features.}
}
\value{
A list of length N, where N is the number of cells tracked for at
least \code{minframes}, with the following entries:
\itemize{
\item{\code{original_ID}: the cell ID from the input file}
\item{\code{cell_missing}: a vector of 1s and 0s with length equal to the
number of frames the cell is tracked over, where a missing frame has a 1
and a non-missing frame 0}
\item{\code{cell_ft}: a matrix of copied features for each frame of each cell}
}
}
\description{
Copy chosen values from an existing table of pre-calculated features.
}
