% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{copyPhaseFeatures}
\alias{copyPhaseFeatures}
\title{Copy cell-frame features from an existing PhaseFocus table}
\usage{
copyPhaseFeatures(file, minframes)
}
\arguments{
\item{file}{The filepath to a CSV file containing features output by PhaseFocus software.}

\item{minframes}{The minimum number of frames a cell must be tracked for to
be included in the output features.}
}
\value{
A dataframe with 1 row corresponding to 1 cell tracked in 1 frame
with the following columns:
\itemize{
\item{\code{frameID}: the numeric frameID}
\item{\code{cellID}: the numeric cellID}
\item{\code{Volume}: a real-valued number}
\item{\code{Sphericity}: a real-valued number}
}
}
\description{
Copies volume and sphericity from a feature table of pre-calculated features as output
by PhaseFocus software. Only the values for the features volume and
sphericity, which require phase information, are copied as all other features
can be calculated by \code{extractFeatures}. Only cells that are tracked for
a minimum of \code{minframes} are included.
}
