% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToCellProfilerAnalyst.R
\name{export_to_cellprofiler_analyst}
\alias{export_to_cellprofiler_analyst}
\title{Export CellPhe frame-specific features to a SQLite database that can be read by CellProfiler Analyst}
\usage{
export_to_cellprofiler_analyst(
  features,
  image_folder,
  image_prefix,
  db_filename = "cellphe.db",
  properties_filename = "cellphe.properties"
)
}
\arguments{
\item{features}{A feature table output by \code{extractFeatures}.}

\item{image_folder}{The folder where the images are stored.}

\item{image_prefix}{The code assumes that images are labelled \if{html}{\out{<prefix>}}_\if{html}{\out{<x>}}.tif, where \code{prefix} is
\code{image_prefix} and \code{x} is the frameID 0-padded to 4 digits.}

\item{db_filename}{The name of the SQLite database to save to.}

\item{properties_filename}{The name of the CellProfiler Analyst properties file to save to.}
}
\value{
Doesn't return anything, but instead creates a SQLite database and a plain text properties file.
}
\description{
CellProfiler Analyst can read the frame-level features from CellPhe for data visualisation
and exploration.
}
