% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract}
\alias{extract}
\title{Extract Features}
\usage{
extract(
  feature_table,
  boundary_coordinates,
  mini_image,
  class_label,
  max_number_of_frames,
  maximum_boundary_length,
  maximum_cell_area,
  cooccurrence_levels,
  number_of_wavelet_levels
)
}
\arguments{
\item{feature_table}{TODO}

\item{boundary_coordinates}{TODO}

\item{mini_image}{TODO}

\item{class_label}{TODO}

\item{max_number_of_frames}{TODO}

\item{maximum_boundary_length}{TODO}

\item{maximum_cell_area}{TODO}

\item{cooccurrence_levels}{TODO}

\item{number_of_wavelet_levels}{TODO}
}
\value{
TODO
}
\description{
Extract Features
}
