% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyHeterogeneousClusters.R
\name{identifyHeterogeneousClusters}
\alias{identifyHeterogeneousClusters}
\title{Identify heterogeneous cell clusters}
\usage{
identifyHeterogeneousClusters(dataset, dataID, k)
}
\arguments{
\item{dataset}{feature table for clustering}

\item{dataID}{a list of cell identifiers (e.g. cell IDs)}

\item{k}{number of desired clusters}
}
\value{
This function plots a dendogram of the obtained hierarchical clustering results, and outputs a matrix where each column \code{n} lists the cells that were assigned cluster \code{n}.
}
\description{
This function performs hierarchical clustering on a given data set in order to identify \code{k} heterogeneous cell clusters, where \code{k} is pre-defined.
}
\examples{
cellclusters<-identifyHeterogeneousClusters(cell1FeatureTable, cellIDs, 3)
}
