% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputBoundaryCoords.R
\name{outputBoundaryCoords}
\alias{outputBoundaryCoords}
\title{Extract a table of cell boundary coordinates from a zip file of ROIs}
\usage{
outputBoundaryCoords(zipFileOfROIs)
}
\arguments{
\item{zipFileOfROIs}{Name of zipped folder containing each ROI for boundary coordinate extraction, note that filenames must be entered in quotation marks and should exclude the .zip extension}
}
\value{
This function outputs a data table of extracted cell boundary coordinates and saves this as a global variable called \code{boundarycoords} where:

The value in \code{boundarycoords[1,1]} is the maximum boundary length of all cells within the data set

\code{boundarycoords[,1]} is a column of frame numbers

\code{boundarycoords[,2]} is a column of cell IDs

\code{boundarycoords[,3]} is a column of boundary lengths

Remaining columns list x and y coordinates alternately until all boundary coordinates have been listed, in which case 0s are used to fill the remainder of the row
}
\description{
Extract a table of cell boundary coordinates from a zip file of ROIs
}
\examples{
outputBoundaryCoords("myZipFolderOfROIs")
}
