% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{prepareBoundaryCoordinates}
\alias{prepareBoundaryCoordinates}
\title{Prepare ROI Boundary Coordinates}
\usage{
prepareBoundaryCoordinates(rois)
}
\arguments{
\item{rois}{A collection of ImageJ ROI data, as a list.}
}
\value{
A collection of simplified ROI data, as a list of integer vectors.
}
\description{
Simplifies a collection of ImageJ ROI data into frame ID, cell ID, and coordinates.
}
