% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{prepareFeatureTable}
\alias{prepareFeatureTable}
\title{Prepare Feature Table}
\usage{
prepareFeatureTable(featureTableFile, cellId)
}
\arguments{
\item{featureTableFile}{File name of a full feature table CSV, as a character vector.}

\item{cellId}{The tracking ID of the cell you want to evaluate, as an integer.}
}
\value{
A matrix containing full feature table data for the passed cell ID.
}
\description{
Reads a full feature table into a matrix, suitable for hand-off to feature extraction.
}
