% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{prepareMiniImage}
\alias{prepareMiniImage}
\title{Prepare Mini Image}
\usage{
prepareMiniImage(rois, frames)
}
\arguments{
\item{rois}{A collection of ImageJ ROIs pertaining to a specific cell, as a list of ijroi objects.}

\item{frames}{A collection of TIFF images, as a list of matrices.}
}
\value{
A collection of frame IDs, widths, heights, and pixel values for each TIFF image, as a list of integer vectors.
}
\description{
Masks a collection of TIFF images to a region of interest for a specific cell.
}
