% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSegmentationTrainingSet.R
\name{prepareSegmentationTrainingSet}
\alias{prepareSegmentationTrainingSet}
\title{Prepare a training set for classification of cell segmentation errors}
\usage{
prepareSegmentationTrainingSet(segerrors, correctsegs)
}
\arguments{
\item{segerrors}{A feature table of ground truth segmentation errors}

\item{correctsegs}{A feature table of ground truth correctly segmented cells}
}
\description{
This function prepares a training set that can be used for classification of cell segmentation errors. This includes use of the
\code{smotefamily::SMOTE()} function to over-sample the minority class to avoid biased classifier training. This function outputs feature tables
for ground truth segmentation errors and correctly segmented cells (with smote data included for the previously under-sampled class), as
well as a list of ground truth labels for each class.
}
