% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{readRois}
\alias{readRois}
\title{Read ROI Files}
\usage{
readRois(directory, cellId)
}
\arguments{
\item{directory}{File name of a directory containing ROI files, as a character vector.}

\item{cellId}{The tracking ID of the cell you want to evaluate, as an integer.}
}
\value{
A list of ROI objects, including additional parameters \code{cellId} and \code{frameId}
}
\description{
Reads a collection of ImageJ ROI files, selecting those relevant to a specific cell ID.
}
