% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{readTiffs}
\alias{readTiffs}
\title{Read a Collection of TIFF images}
\usage{
readTiffs(directory)
}
\arguments{
\item{directory}{File name of a directory containing TIFF files, as a character vector.}
}
\value{
A collection of TIFF images, as a list of matrices.
}
\description{
Reads a directory of TIFF images, sorting by sequence number in the file name.
}
