% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePredictedSegErrors.R
\name{removePredictedSegErrors}
\alias{removePredictedSegErrors}
\title{Remove predicted segmentation errors from test set}
\usage{
removePredictedSegErrors(testset, k, predictedSegErrors)
}
\arguments{
\item{testset}{Test set for segmentation error predictions to be made}

\item{k}{Column index for the column of cell identifiers within the test set, e.g. if column 2 of the test set is a column of cell identifiers then here k = 2}

\item{predictedSegErrors}{Output from either predictSegErrors() or predictSegErrors_Ensemble(), a list of cell identifiers for cells classified as segmentation error}
}
\description{
This function can be used to automate removal of predicted segmentation errors from the test set
}
