% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{simplifyRoi}
\alias{simplifyRoi}
\title{Simplify an ROI Object}
\usage{
simplifyRoi(roi)
}
\arguments{
\item{roi}{ImageJ ROI data, as an ijroi object.}
}
\value{
ROI frame ID, cell ID, and coordinates, as an integer vector.
}
\description{
Simplifies an ROI object to its frame ID, cell ID, and region of interest coordinates.
}
