% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBySeparationThreshold.R
\name{subsetBySeparationThreshold}
\alias{subsetBySeparationThreshold}
\title{Only include features above a set separation threshold}
\usage{
subsetBySeparationThreshold(outputfile, separationscores, t)
}
\arguments{
\item{outputfile}{Feature table to be subset}

\item{separationscores}{list of separation scores, each element with a different separation threshold}

\item{t}{which index from separation scores to use}
}
\description{
Subsets a data file to only include features above a set separation threshold
}
