% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsFromTimeSeries.R
\name{varsFromTimeSeries}
\alias{varsFromTimeSeries}
\title{Calculate variables from the time series of each extracted cell feature.}
\usage{
varsFromTimeSeries(df)
}
\arguments{
\item{df}{A data frame, the same as the output from \code{extractFeatures}.}
}
\value{
A data frame with cells in rows and variables describing their behaviour over time in columns.
}
\description{
Reads in the time series for any pre-existing features as well as the
output of \code{extractFeatures}. Variables are calculated from the time series,
providing both summary statistics and indicators of time-series behaviour
at different levels of detail obtained via wavelet analysis.
}
