% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSeparationScores.R
\name{calculateSeparationScores}
\alias{calculateSeparationScores}
\title{Calculate a table containing separation scores for each variable}
\usage{
calculateSeparationScores(group1data, group2data, threshold)
}
\arguments{
\item{group1data}{Output file of variables for group 1}

\item{group2data}{Output file of variables for group 2}

\item{threshold}{Separation threshold, only features achieving separation greater than or equal to this threshold will be output}
}
\value{
A data frame of separation values where column 1 is a list of variable indices, column 2 a list of variable names, and column 3 the corresponding separation scores
}
\description{
This function can be used to calculate the separation score for each variable, where a feature selection threshold can be defined by the user to identify discriminatory variables
}
\examples{
separationscores<-calculateSeparationScores(Untreated, Treated, 0.1)
}
