% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyFeatures.R
\name{copyFeatures}
\alias{copyFeatures}
\title{Copy metadata and cell-frame features from an existing PhaseFocus or Trackmate table}
\usage{
copyFeatures(file, minframes, source = c("Phase", "Trackmate"))
}
\arguments{
\item{file}{The filepath to a CSV file containing features output by PhaseFocus or Trackmate software.}

\item{minframes}{The minimum number of frames a cell must be tracked for to
be included in the output features.}

\item{source}{The name of the software that produced the metadata file, either 'Phase' or 'Trackmate'
are currently supported.}
}
\value{
A dataframe with 1 row corresponding to 1 cell tracked in 1 frame
with the following columns:
\itemize{
\item{\code{FrameID}: the numeric FrameID}
\item{\code{CellID}: the numeric CellID}
\item{\code{ROI_filename}: the label used to refer to this ROI}
\item{\code{Volume}: a real-valued number}
\item{\code{Sphericity}: a real-valued number}
}
}
\description{
Loads the frame and cell IDs along with the filename used to refer to each ROI.
For PhaseFocus generated data, volume and sphericity features are also extracted.
Only cells that are tracked for a minimum of \code{minframes} are included.
}
