% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{copyPhaseFeatures}
\alias{copyPhaseFeatures}
\title{Copy features from an existing PhaseFocus feature table}
\usage{
copyPhaseFeatures(file, minframes)
}
\arguments{
\item{file}{The filepath to a CSV file containing features output by PhaseFocus software.}

\item{minframes}{The minimum number of frames a cell must be tracked for to
be included in the output features.}
}
\value{
A list of length N, where N is the number of cells tracked for at
least \code{minframes}, with the following entries:
\itemize{
\item{\code{original_ID}: the cell ID from the input file}
\item{\code{cell_missing}: a vector of 1s and 0s with length equal to the
number of frames the cell is tracked over, where a missing frame has a 1
and a non-missing frame 0}
\item{\code{cell_ft}: a matrix of copied features for each frame of each cell}
}
}
\description{
Copy chosen values from a feature table of pre-calculated features as output
by PhaseFocus software. Only the values for the features volume and
sphericity, which require phase information, are copied as all other features
can be calculated by \code{extractFeatures}. Only cells that are tracked for
a minimum of \code{minframes} are copied into the new feature table.
}
