% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Calculates cell features from timelapse videos}
\usage{
extractFeatures(df, roi_folder, frame_folder, framerate = 1)
}
\arguments{
\item{df}{DataFrame where every row corresponds to a combination of a cell
tracked in a frame. It must have at least columns \code{CellID}, \code{FrameID} and
\code{ROI_filename} along with any additional features.}

\item{roi_folder}{A path to a directory containing multiple Report Object Instance
(ROI) files named in the format \code{cellid}-\code{frameid}.roi}

\item{frame_folder}{A path to a directory containing multiple frames in TIFF format.
It is assumed these are named under the pattern \code{<experiment name>-<frameid>.tif}, where
\code{<frameid>} is a 4 digit zero-padded integer.}
}
\value{
A dataframe with 76 columns and 1 row per cell per frame it's present in:
\itemize{
\item{\code{frameID}: the numeric frameID}
\item{\code{cellID}: the numeric cellID}
\item{\code{ROI_filename}: the ROI filename}
\item{\code{...}: 74 frame specific features}
\item{\code{...}: Any other data columns that were present in \code{df}}
}
}
\description{
Calculates 74 features related to size, shape, texture and movement for each
cell on every non-missing frame, as well as the cell density around each
cell on each frame.
NB: while the ROI filenames are expected to be provided in \code{df} and found in
\code{roi_folder}, the frame filenames are just expected to follow the naming convention
\code{<some text>-<frameID>.tiff}, where frameID is a 4 digit leading zero-padded number, corresponding
to the \code{frameID} column in \code{df}.
}
