% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Calculates cell features from timelapse videos}
\usage{
extractFeatures(df, frames, roi_folder, min_frames, framerate = 1)
}
\arguments{
\item{df}{DataFrame where every row corresponds to a combination of a cell
tracked in a frame. It must have at least columns \code{CellID} and \code{FrameID},
along with any additional features.}

\item{frames}{The frames themselves, a list of TIFF images.}

\item{roi_folder}{A path to a directory containing multiple Report Object Instance
(ROI) files named in the format \code{cellid}-\code{frameid}.roi}

\item{min_frames}{The minimum number of frames a cell must be tracked for to
be included in the output features.}
}
\value{
A dataframe with 76 columns and 1 row per cell per frame it's present in:
\itemize{
\item{\code{frameID}: the numeric frameID}
\item{\code{cellID}: the numeric cellID}
\item{\code{...}: 72 frame specific features
\item{\code{xcentres}: the x-coordinate of the cell in that frame}
\item{\code{ycentres}: the y-coordinate of the cell in that frame}
}
}
\description{
Calculates 1109 features related to size, shape, texture and movement for each
cell on every non-missing frame, as well as the cell density around each
cell on each frame and a measure describing the trajectory of the cell
over all frames.
}
\details{
Reads in information on cell boundaries from ROI files, the cell identifiers
from the original tracking (i.e. before cells tracked for less than
min_frames were removed), original_IDs, information on missing frames for
each cell in a list of vectors with missing frames indicated by 1 and
non-missing frames by 0, missing_frames, and the normalised images for every frame, normalised_frames.
}
