% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{extractFeatures}
\alias{extractFeatures}
\title{TODO}
\usage{
extractFeatures(
  file,
  original_IDs,
  missing_frames,
  normalised_frames,
  min_frames
)
}
\arguments{
\item{file}{TODO}

\item{original_IDs}{TODO}

\item{missing_frames}{TODO}

\item{normalised_frames}{TODO}

\item{min_frames}{TODO}
}
\value{
TODO
}
\description{
TODO
}
