% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSegErrors.R
\name{predictSegErrors}
\alias{predictSegErrors}
\title{Train and test decision trees for segmentation error prediction}
\usage{
predictSegErrors(
  segerrors,
  correctsegs,
  num,
  testset,
  dataID,
  proportion,
  dup_size = 1
)
}
\arguments{
\item{segerrors}{A feature table of ground truth segmentation errors}

\item{correctsegs}{A feature table of ground truth correctly segmented cells}

\item{num}{Number of decision trees to be trained}

\item{testset}{Test set for segmentation error predictions to be made}

\item{dataID}{List of test set identifiers (e.g. cell IDs)}

\item{proportion}{Proportion of votes needed for a final classification of segmentation error to be made (e.g. 0.7 if 70\% of the votes are needed for segmentation error classification to be made)}

\item{dup_size}{Passed to smotefamily::SMOTE. The number of times to
duplicate the minority class.}
}
\value{
This function returns the list of identifiers that were predicted as segmentation errors
}
\description{
This function trains \code{num} decision trees and uses them to predict whether or not new cells experience segmentation errors. Final classifications are
made via a voting system, where a cell is classified as segmentation error if more than a defined \code{proportion} of decision trees predict it as such. This code outputs a list
of cells that were classified as segmentation error.
}
