% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsFromTimeSeries.R
\name{varsFromTimeSeries}
\alias{varsFromTimeSeries}
\title{TODO}
\usage{
varsFromTimeSeries(features, new_features, expname, originalIDs)
}
\arguments{
\item{features}{TODO}

\item{new_features}{TODO}

\item{expname}{TODO}

\item{originalIDS}{TODO}

\item{return}{TODO}
}
\description{
TODO
}
